backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary
python-decouple
dj-database-url
supertokens-python
typing 
django-cors-headers
djangorestframework
django-filter
markdown
